暂时提交

+++ src/main/java/com/pe/exchange/dao/UserBalanceDao.java
@@ -6,12 +6,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
 public interface UserBalanceDao extends JpaRepository<UserBalance,Integer> {
 
-    Integer findBanalceByUserId(Integer userId);
+    BigDecimal findBanalceByUserId(Integer userId);
     @Query(value = "update user_balance set balance=balance+:amount where user_id=:userId",nativeQuery = true)
-    int addBalance(@Param("userId") Integer userId,@Param("amount")Integer amount);
+    int addBalance(@Param("userId") Integer userId,@Param("amount")BigDecimal amount);
     @Query(value = "update user_balance set balance=balance-:amount where user_id=:userId",nativeQuery = true)
-    int subBalance(@Param("userId") Integer userId,@Param("amount")Integer amount);
+    int subBalance(@Param("userId") Integer userId,@Param("amount")BigDecimal amount);
 }
